Testing: Refactor software tests into dedicated directory `tests`
git mv src/crate/client/test* tests/client/
git mv src/crate/testing/test* tests/testing/
