+++ tests/assets/mappings/locations.sql
