+++ tests/assets/pki/client_invalid.pem
