+++ tests/client/test_exceptions.py
