+++ tox.ini
@@ -11,4 +11,4 @@ deps =
     mock
     urllib3
 commands =
-    zope-testrunner -c --test-path=src
+    zope-testrunner -c --path=tests
